requirements.txt
rplcd
python-vlc
requests
RPi.GPIO
tinydb
pyyaml